@@ -1,15 +1,17 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="5.5.0" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.2.3" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="../../Source/SDK/SDK.csproj" />
+    <ProjectReference Include="../../Source/Extensions.AspNet/Extensions.AspNet.csproj" />
+    <ProjectReference Include="../../Source/Extensions.Discovery/Extensions.Discovery.csproj" />
   </ItemGroup>
 
 </Project>
